requirements.txt
# ================================================================
# SECEL — Dépendances Python
# Compatible Python 3.10+
# ================================================================

# Framework web
Flask>=3.1.0
Werkzeug>=3.1.0

# ORM base de données
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.48

# Authentification et sessions
Flask-Login>=0.6.3

# Formulaires et protection CSRF
Flask-WTF>=1.2.0
WTForms>=3.2.0

# Rate limiting
Flask-Limiter>=4.1.1

# Tâches planifiées (auto-suppression — désactivé sur Vercel)
APScheduler>=3.10.4

# Variables d'environnement (.env)
python-dotenv>=1.0.0

# OAuth 2.0 (Google & LinkedIn)
authlib>=1.3.0
requests>=2.31.0

# Driver PostgreSQL pour SQLAlchemy (Supabase)
psycopg2-binary>=2.9.9

# Client Python Supabase
supabase>=2.4.0

# Google Gemini AI — chatbot SELA
google-genai>=1.0.0
